Fixes, make test sequential
