+++ tests/test_hyperparams_opt.py
@@ -2,7 +2,9 @@ import glob
 import os
 import subprocess
 
+import optuna
 import pytest
+from optuna.trial import TrialState
 
 
 def _assert_eq(left, right):
@@ -120,3 +122,55 @@ def test_optimize_log_path(tmp_path):
     ]
     return_code = subprocess.call(["python", "scripts/parse_study.py"] + args)
     _assert_eq(return_code, 0)
+
+
+def test_multiple_workers(tmp_path):
+    study_name = "test-study"
+    storage = f"sqlite:///{tmp_path}/optuna.db"
+    # n trials per worker
+    n_trials = 2
+    # max total trials
+    max_trials = 3
+    # 1st worker will do 2 trials
+    # 2nd worker will do 1 trial
+    # 3rd worker will do nothing
+    n_workers = 3
+    args = [
+        "-optimize",
+        "--no-optim-plots",
+        "--storage",
+        storage,
+        "--n-trials",
+        str(n_trials),
+        "--max-total-trials",
+        str(max_trials),
+        "--study-name",
+        study_name,
+        "--n-evaluations",
+        str(1),
+        "-n",
+        str(100),
+        "--algo",
+        "a2c",
+        "--env",
+        "Pendulum-v1",
+        "--log-folder",
+        tmp_path,
+        "-params",
+        "n_envs:1",
+    ]
+
+    # Sequencial execution to avoid race conditions
+    workers = []
+    for _ in range(n_workers):
+        worker = subprocess.Popen(
+            ["python", "train.py"] + args, stdout=subprocess.PIPE, stderr=subprocess.PIPE, universal_newlines=True
+        )
+        worker.wait()
+        workers.append(worker)
+
+    study = optuna.load_study(study_name=study_name, storage=storage)
+    assert len(study.get_trials(states=(TrialState.COMPLETE, TrialState.PRUNED))) == max_trials
+
+    for worker in workers:
+        assert worker.returncode == 0, "STDOUT:\n{}\nSTDERR:\n{}\n".format(*worker.communicate())
